Convert helper methods to static and consolidate replicator logic.
